refactor(metrics): split internal/metrics into per-feature bundles (TD-6)
Splits internal/metrics/metrics.go (1171 LOC) into a slim core plus
12 per-feature files: proxy, websocket, dlp, session, tls, airlock,
cross_request, scan_api, kill_switch, shield, capture, stats_handler.

The Metrics struct definition stays in metrics.go and New() dispatches
to a register*Metrics(reg) helper per file. Each helper allocates
that feature's prometheus handles, registers them on the per-Metrics
registry, and stores them on the shared struct. No metric name, label
set, help string, or bucket array is changed; no public method or
field signature is changed; no caller is touched.

Verified:
- 57 prometheus.New(Counter|Gauge|Histogram)(Vec)? sites preserved
- 62 *Metrics method signatures byte-identical (sorted diff)
- internal/metrics tests pass unchanged
- Full repo go test -race -count=1 ./... green
- golangci-lint v2.10.1 ./... clean
- 3-layer security suite (in-repo Go, private white-box, 885 black-box
  pen test assertions) all pass

Closes TD-6 (v2.4 phase 1).
